requirements.txt
python-dotenv==1.0.1
openai==1.37.1
Pillow==10.4.0
numpy==2.1.1
requests==2.32.3
scikit-learn==1.5.1
nltk==3.8.1
requests~=2.32.3
newspaper3k==0.2.8
lxml_html_clean==0.2.0